% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dt_create}
\alias{dt_create}
\title{Create a dt object}
\usage{
dt_create(data, group, stat_num = "median", digits = 2,
  spec_var = NULL, keep_missing = F, compare = T, overall = F)
}
\arguments{
\item{data}{A data frame which contains only factor or numeric variables.}

\item{group}{The name of the grouping variable.}

\item{stat_num}{Characters. Specify which statistics should be provide.}

\item{digits}{Integer. Specify the number of digits to display.}

\item{spec_var}{List. A named list of the variables for which you want
specify row name or number of digits to display. Each variable-named
element need to be a named list (as name and digit). If provides here, the
specific digit argument overwrite the digits argument. The order of the
variable-named element define the order of rows in the table.}

\item{keep_missing}{Logical. If FALSE (default) observations which have a
missing value for the grouping variable are exclude to the dataset.
Otherwise an extra group named "Missing data" will be add to the table to
describe this observations.}

\item{compare}{Logical. If TRUE (default) a statistical comparison is added
to the table.}

\item{overall}{Logical. If TRUE an extra column with the statistics of the
entire dataset will be added to the table.}
}
\value{
A dt object which is a list
}
\description{
Create a dt object
}
