% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr.R
\name{dt_save}
\alias{dt_save}
\title{Save a rendered dt to pdf, html or docx}
\usage{
dt_save(x, file)
}
\arguments{
\item{x}{A table rendered by one of those functions \code{dt_to_html()}, \code{dt_to_latex()} or \code{dt_to_flextable()}}

\item{file}{A file name without extension specified.}

\item{path}{A path to directory where file will be save. If NULL file will be save in the working directory.}
}
\value{
A pdf, html or docx file
}
\description{
Save a rendered dt to pdf, html or docx
}
