% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr.R
\docType{package}
\name{dtr}
\alias{dtr}
\alias{dtr-package}
\title{dtr: Descriptive Table R}
\description{
The purpose of the package is to quickly get a numerical description of a set
of variables as well a statistical comparison between groups of interest. This
is an usual task in medical researches as clinical trials or epidemiological
studies to better know the population, but reporting it may become tedious.
}
\section{Create}{
 The main function \code{\link{dt_create}} takes a dataset, a
 grouping variable and additionnal arguments to personnalized the output
 (e.g. number of digits, statistical comparisons, overall description).
 The function return a dt object.
}

\section{Inspect}{
 Original descriptive table as characters data frame can be
 extract from the dt object. Details of each statistical test can be extract ...
}

\section{Rendering to report}{
 dt objects from \code{\link{dt_create}} can be pass
 to \itemize{ \item \code{\link{dt_to_html}} or \code{\link{dt_to_latex}} to
 get the corresponding HTML or LaTeX code generated by the
 \code{\link[kableExtra:kableExtra]{kableExtra}} package \item
 \code{\link{dt_to_flextable}} to get the corresponding flextable object
 generated by the \code{\link[flextable:flextable-package]{flextable}}
 package \item \code{\link{print}} method to get an ASCII version of the
 table } Output returned by this 3 functions can be directly use in Rmarkdown
 documents
}

\section{Export}{
 Instead of use rendering functions directly in Rmarkdown
 documents, tables outputed from \code{\link{dt_to_html}},
 \code{\link{dt_to_latex}} or \code{\link{dt_to_flextable}} can be save as
 \strong{.html}, \strong{.pdf} or \strong{.docx} files with the
 \code{\link{dt_save}} function
}

