% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitable.R
\name{epitable}
\alias{epitable}
\title{Create an epitable object}
\usage{
epitable(data, group, stat.num = "median", digits = 2,
  spec.var = NULL, keep.missing = F, compare = T, overall = F)
}
\arguments{
\item{data}{A data frame which contains only factor or numeric variables.}

\item{group}{Characters. A grouping variable name.}

\item{stat.num}{Characters. Specify which statistics should be provide.}

\item{digits}{Integer. Specify the number of digits to display.}

\item{spec.var}{A list. Named list of the variables for which it specifies name and digits to display.}

\item{keep.missing}{A T/F value. Indicate if statistics should be provide for cases with missing grouping variable.}

\item{compare}{A T/F value. Indicate if a statistical comparison should be added.}

\item{overall}{A T/F value. Indicate if statistics should be provide for the entire dataset.}
}
\value{
An epitable object
}
\description{
Create an epitable object
}
