% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitable_to_html.R
\name{epitable_to_html}
\alias{epitable_to_html}
\title{Render an epitable object in html format}
\usage{
epitable_to_html(object, title = NULL, font_size = NULL, save = F)
}
\arguments{
\item{object}{An epitable object returned by the \code{epitable()} function}

\item{title}{A character string to add a title to the table.}

\item{font_size}{A numeric value to set the font size.}

\item{save}{A T/F logical. If true it will save the table within a html page
in the working directories.}
}
\description{
Render an epitable object in html format
}
\examples{
data <- ggplot2::mpg \%>\%
   select(manufacturer, displ, cyl, cty, hwy) \%>\%
   filter(manufacturer \%in\% c("audi", "chevrolet", "dodge", "ford"))

data \%>\% epitable(group = cyl) \%>\%
   epitable_to_html(title = "Compare characteristics of manufacturer", save = T)
}
