\name{LDTdata}
\docType{data}
\alias{LDTdata}
\title{
LDT data set
}
\description{
This data set was generates from sequentially randomized clinical trials as described in the simulation study of Lunceford, Davidian and Tsiatis (2002). It contains the following variables: "X" is the first-stage indicator, X=0 if assigned to A1, and X=1 if assigned to A2; "R" is the response status, R=1 for responders, and R=0 for non-responders; "Z" is the second-stage indicator among responders (R=1), Z=0 if assigned to B1, and Z=1 if assigned to B2; "U" is the observed survival time, U is death time if delta=1, and U is censoring time if delta=0; "delta" is the censoring indicator, delta=1 for event, and delta=0 for censored.
}
\usage{
LDTData
}
\format{
A data frame with rows corresponding to patients.
}
\references{
Lunceford JK, Davidian M, Tsiatis AA: Estimation of survival distributions of treatment policies in two-stage randomization designs in clinical trials. Biometrics 58:48-57, 2002
}
\source{
Generated by Xinyu Tang in R
}
\examples{
\dontrun{
data("LDTdata")
}
}
\keyword{datasets}