\name{WRSEdata}
\docType{data}
\alias{WRSEdata}
\title{
WRSE data set
}
\description{
This data set was generates from sequentially randomized clinical trials as described in the simulation study of Guo and Tsiatis (2005). It contains the following variables: "X" is the first-stage indicator, X=0 if assigned to A1, and X=1 if assigned to A2; "TR" is the time to response for responders (R=1); "R" is the response status, R=1 for responders, and R=0 for non-responders; "Z" is the second-stage indicator among responders (R=1), Z=0 if assigned to B1, and Z=1 if assigned to B2; "U" is the observed survival time, U is death time if delta=1, and U is censoring time if delta=0; and "delta" is the censoring indicator, delta=1 for event, and delta=0 for censored.
}
\usage{
WRSEdata
}
\format{
A data frame with rows corresponding to patients.
}
\references{
Guo X, Tsiatis AA: A weighted risk set estimator for survival distributions in two-stage randomization designs with censored survival data. Int. J. Biostatistics 1:1-15, 2005
}
\source{
Generated by Xinyu Tang in R
}
\examples{
\dontrun{
data("WRSEdata")}
}
\keyword{datasets}