% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr.R
\name{dt_create}
\alias{dt_create}
\title{Create an dt object}
\usage{
dt_create(data, group, stat_num = "median", digits = 2,
  spec_var = NULL, keep_missing = F, compare = T, overall = F)
}
\arguments{
\item{data}{A data frame which contains only factor or numeric variables.}

\item{group}{Characters. A grouping variable name.}

\item{stat_num}{Characters. Specify which statistics should be provide.}

\item{digits}{Integer. Specify the number of digits to display.}

\item{spec_var}{A list. Named list of the variables for which it specifies name and digits to display.}

\item{keep_missing}{A T/F value. Indicate if statistics should be provide for cases with missing grouping variable.}

\item{compare}{A T/F value. Indicate if a statistical comparison should be added.}

\item{overall}{A T/F value. Indicate if statistics should be provide for the entire dataset.}
}
\value{
An epitable object
}
\description{
Create an dt object
}
