% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_render.R
\name{dt_to_flextable}
\alias{dt_to_flextable}
\title{Render an epitable x in a word document}
\usage{
dt_to_flextable(x, title = NULL, font_size = NULL)
}
\arguments{
\item{x}{An epitable x returned by the \code{epitable()} function}

\item{title}{A character string to add a title to the table.}

\item{font_size}{A numeric value to set the font size.}
}
\description{
Render an epitable x in a word document
}
\examples{
data <- ggplot2::mpg \%>\%
   select(manufacturer, displ, cyl, cty, hwy) \%>\%
   filter(manufacturer \%in\% c("audi", "chevrolet", "dodge", "ford"))

data \%>\% epitable(group = cyl) \%>\%
   epitable_to_word(title = "Compare characteristics of manufacturer")
}
