% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{dt_to_latex}
\alias{dt_to_latex}
\title{Render an dt object in latex format}
\usage{
dt_to_latex(x, title = NULL, font_size = NULL)
}
\arguments{
\item{x}{A dt object returned by the \code{create_dt()} function}

\item{title}{A title to the table.}

\item{font_size}{A numeric value to set the font size.}
}
\description{
Render an dt object in latex format
}
\examples{
data <- ggplot2::mpg \%>\%
   select(manufacturer, displ, cyl, cty, hwy) \%>\%
   filter(manufacturer \%in\% c("audi", "chevrolet", "dodge", "ford")) \%>\%

data \%>\% epitable(group = cyl) \%>\%
   epitable_to_latex(title = "Compare characteristics of manufacturer", save = T)
}
